requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask==3.1.0
pandas==2.2.3
duckdb==1.2.1
llama-cpp-python==0.2.85
pyarrow==19.0.1
pylance==0.23.2
huggingface-hub